feat(TAT-125): split methodology details code into dynamic components
